requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
openai>=1.50.0
streamlit>=1.35.0
python-dotenv>=1.0.0